# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "novamem",
    "version": "1.1.7",
    "description": "Tiered memory service with hybrid search (keyword + vector + graph), per-user isolation, and project-scoped sub-brains.\n\n**Auth**: every authenticated route accepts either an `nm_…` user bearer (created from the dashboard's API Tokens page) or a Better Auth session cookie/bearer (set by `POST /api/auth/sign-in/email`). Both flow through the same auth hook and resolve to the same user. Better Auth's own routes are mounted at `/api/auth/*` — see the Better Auth documentation for their shapes.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "nm_…",
        "description": "Per-device API token belonging to one user. Carries every right the owning user has — data plane, project CRUD, membership, metrics. Created from the dashboard's API Tokens page; the plaintext is shown once."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nm.session_token",
        "description": "Better Auth session cookie set by /api/auth/sign-in/email. HttpOnly + SameSite=Lax. Browsers attach it automatically."
      }
    },
...
  },
  "paths": {
...
    "/v1/admin/tokens/revoke": {
      "post": {
        "summary": "Revoke a user bearer by plaintext token",
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/users": {
      "post": {
        "summary": "Provision a user (and optionally a bearer token) non-interactively",
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 256
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "tokenLabel": {
                    "type": "string",
                    "maxLength": 128
                  },
                  "tokenScope": {
                    "type": "string",
                    "enum": [
                      "full",
                      "read_only"
                    ]
                  },
                  "tokenExpiresInDays": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 3650
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "summary": "List every user with entry + token counts",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/users/{id}": {
      "delete": {
        "summary": "Delete a user and everything they own (irreversible)",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "dryRun",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/users/{id}/quota": {
      "put": {
        "summary": "Set per-user quota overrides (null clears back to server default)",
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxEntries": {
                    "nullable": true,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "writesPerMinute": {
                    "nullable": true,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/audit-log": {
      "get": {
        "summary": "Tail the audit log",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "default": 200,
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 500
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "this server"
    }
  ],
...
}